refactor: unify type system and eliminate duplicates
- Remove duplicate CoreSecurityConfig from core/server/types.ts
- Remove duplicate directive types from nitro/types.ts, import from
  core/types/define.ts instead
- Move StandardSchemaV1 to dedicated file (core/types/standard-schema.ts)
- Replace GraphQLArgumentType 35-literal union with template literal type
- Fix DEFAULT_PATHS_CONFIG null values to undefined for type safety
